+++ src/main/kotlin/uk/gov/justice/digital/hmpps/dprdpdsservice/HmppsDprDpdsService.kt
@@ -3,10 +3,11 @@ package uk.gov.justice.digital.hmpps.dprdpdsservice
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 
-@SpringBootApplication(exclude = [
-  org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration::class,
-    org.springframework.boot.actuate.autoconfigure.security.servlet.ManagementWebSecurityAutoConfiguration::class
-]
+@SpringBootApplication(
+  exclude = [
+    org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration::class,
+    org.springframework.boot.actuate.autoconfigure.security.servlet.ManagementWebSecurityAutoConfiguration::class,
+  ],
 )
 class HmppsDprDpdsService
 fun main(args: Array<String>) {
