+++ src/main/kotlin/uk/gov/justice/digital/hmpps/dprdpdsservice/controller/DefinitionsController.kt
@@ -5,18 +5,17 @@ import io.swagger.v3.oas.annotations.tags.Tag
 import jakarta.servlet.http.HttpServletRequest
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RestController
-import uk.gov.justice.digital.hmpps.digitalprisonreportinglib.data.model.ProductDefinition
-import uk.gov.justice.digital.hmpps.dprdpdsservice.service.DefinitionsService
+import uk.gov.justice.digital.hmpps.dprdpdsservice.repository.DefinitionsRepository
 
 @RestController
 @Tag(name = "Data Product Definitions API")
-class DefinitionsController(val definitionsService: DefinitionsService) {
+class DefinitionsController(val definitionsRepository: DefinitionsRepository) {
   @Operation(
     description = "Gets all the definitions by directory",
   )
   @GetMapping("/definitions/**")
-  fun getDefinitions(request: HttpServletRequest): List<Map<String,Any>> {
+  fun getDefinitions(request: HttpServletRequest): List<Map<String, Any>> {
     val path: String = request.requestURI
-    return definitionsService.getDefinitions(path.removePrefix("/"))
+    return definitionsRepository.getDefinitions(path.removePrefix("/"))
   }
 }
