+++ src/main/kotlin/uk/gov/justice/digital/hmpps/dprdpdsservice/repository/DefinitionsRepository.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.dprdpdsservice.service
+package uk.gov.justice.digital.hmpps.dprdpdsservice.repository
 
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
@@ -10,23 +10,22 @@ import java.nio.file.Paths
 import kotlin.io.path.name
 
 @Service
-class DefinitionsService() {
+class DefinitionsRepository() {
 
-  fun getDefinitions(path: String): List<Map<String,Any>> {
-   val gson : Gson = GsonBuilder()
+  fun getDefinitions(path: String): List<Map<String, Any>> {
+    val gson: Gson = GsonBuilder()
       .create()
     val pathURI = this::class.java.classLoader.getResource(path)?.toURI() ?: return emptyList()
     val dirs = Files.walk(Paths.get(pathURI), 1)
       .filter { Files.isRegularFile(it) }
-      .filter { item -> item.toString().endsWith(".json")  }
+      .filter { item -> item.toString().endsWith(".json") }
       .toList()
-    val result = dirs.map<Path, Map<String,Any>> {
+    val result = dirs.map<Path, Map<String, Any>> {
       gson.fromJson(
         this::class.java.classLoader.getResource("$path/${it.name}")?.readText(),
-        object : TypeToken<Map<String,Any>>() {}.type
+        object : TypeToken<Map<String, Any>>() {}.type,
       )
     }
     return result
   }
-
 }
